refactor: use common version cmd
